pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-spectra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "einops>=0.8.1",
    "flask>=2.3.3",
    "pillow>=11.2.1",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.0",
    "werkzeug>=2.3.7",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-mock",
]
examples = [
    "matplotlib>=3.7",
    "tqdm>=4.65",
]
dev = [
    "jupyter>=1.1.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "manual: marks tests requiring human verification (deselect with '-m \"not manual\"')",
    "requires_capability(name): marks tests requiring specific model capability",
    "slow: marks slow tests",
]
